Add parser tests
