+++ src/parser/mod.rs
@@ -1,17 +1,14 @@
 mod common;
 mod filter;
 mod path;
+mod record;
 mod string;
 
 pub use common::ws;
 pub use filter::parse_filter;
 pub use path::{parse_path, parse_path_list};
-pub use string::parse_string;
-
-// OLD
-mod record;
-
 pub use record::{
     parse_field, parse_field_occurrence, parse_field_tag, parse_record,
     parse_subfield, parse_subfield_code,
 };
+pub use string::parse_string;
